refactor(Option): split out operations, instances and dsls
