+++ packages/core/src/Option/dsl/gen.ts
@@ -1,0 +1,6 @@
+// ets_tracing: off
+
+import * as P from "../../Prelude"
+import { Monad } from "../instances/Monad"
+
+export const gen = P.genF(Monad)
