+++ packages/core/src/Option/dsl/matchers.ts
@@ -1,0 +1,10 @@
+// ets_tracing: off
+
+import * as P from "../../Prelude"
+import { Covariant } from "../instances/Covariant"
+
+/**
+ * Matchers
+ */
+export const { match, matchIn, matchMorph, matchTag, matchTagIn } =
+  P.matchers(Covariant)
