+++ packages/core/src/Option/dsl/struct.ts
@@ -1,0 +1,7 @@
+// ets_tracing: off
+
+import * as P from "../../Prelude"
+import { Applicative } from "../instances/Applicative"
+import { Monad } from "../instances/Monad"
+
+export const struct = P.structF({ ...Monad, ...Applicative })
