+++ packages/core/src/Option/dsl/tuple.ts
@@ -1,0 +1,7 @@
+// ets_tracing: off
+
+import * as P from "../../Prelude"
+import { Applicative } from "../instances/Applicative"
+import { Monad } from "../instances/Monad"
+
+export const tuple = P.tupleF({ ...Monad, ...Applicative })
