+++ packages/core/src/Option/instances/Any.ts
@@ -1,0 +1,10 @@
+// ets_tracing: off
+
+import * as O from "@effect-ts/system/Option"
+
+import type { OptionURI } from "../../Modules"
+import * as P from "../../Prelude"
+
+export const Any = P.instance<P.Any<[P.URI<OptionURI>]>>({
+  any: () => O.some({})
+})
