+++ packages/core/src/Option/instances/Witherable.ts
@@ -1,0 +1,9 @@
+// ets_tracing: off
+
+import type { OptionURI } from "../../Modules"
+import * as P from "../../Prelude"
+import { compactF } from "../operations/compactF"
+
+export const Witherable = P.instance<P.Witherable<[P.URI<OptionURI>]>>({
+  compactF
+})
