+++ packages/core/src/Option/instances/index.ts
@@ -1,0 +1,21 @@
+// ets_tracing: off
+
+// codegen:start {preset: barrel, include: ./*.ts}
+export * from "./Any"
+export * from "./Applicative"
+export * from "./AssociativeBoth"
+export * from "./AssociativeEither"
+export * from "./AssociativeFlatten"
+export * from "./Compactable"
+export * from "./Covariant"
+export * from "./Extend"
+export * from "./Fail"
+export * from "./Filterable"
+export * from "./Foldable"
+export * from "./ForEach"
+export * from "./IdentityBoth"
+export * from "./IdentityFlatten"
+export * from "./Monad"
+export * from "./Wiltable"
+export * from "./Witherable"
+// codegen:end
