+++ packages/core/src/Option/operations/alt.ts
@@ -1,0 +1,7 @@
+// ets_tracing: off
+
+import * as P from "../../Prelude"
+import { AssociativeEither } from "../instances/AssociativeEither"
+import { Covariant } from "../instances/Covariant"
+
+export const alt = P.orElseF({ ...Covariant, ...AssociativeEither })
