+++ packages/core/src/Option/operations/getFirstAssociative.ts
@@ -1,0 +1,20 @@
+// ets_tracing: off
+
+import * as O from "@effect-ts/system/Option"
+
+import type { Associative } from "../../Associative"
+import { makeAssociative } from "../../Associative"
+
+/**
+ * `Associative` returning the left-most non-`None` value
+ *
+ * | x       | y       | combine(y)(x) |
+ * | ------- | ------- | ------------- |
+ * | none    | none    | none          |
+ * | some(a) | none    | some(a)       |
+ * | none    | some(a) | some(a)       |
+ * | some(a) | some(b) | some(a)       |
+ */
+export function getFirstAssociative<A>(): Associative<O.Option<A>> {
+  return makeAssociative((x, y) => (O.isNone(x) ? y : x))
+}
