+++ packages/core/src/Option/operations/index.ts
@@ -1,0 +1,29 @@
+// ets_tracing: off
+
+export * from "@effect-ts/system/Option"
+
+// codegen:start {preset: barrel, include: ./*.ts}
+export * from "./alt"
+export * from "./compactF"
+export * from "./elem"
+export * from "./filter"
+export * from "./filterMap"
+export * from "./forEachF"
+export * from "./getApplyAssociative"
+export * from "./getApplyIdentity"
+export * from "./getEqual"
+export * from "./getFirst"
+export * from "./getFirstAssociative"
+export * from "./getFirstIdentity"
+export * from "./getIdentity"
+export * from "./getLast"
+export * from "./getLastAssociative"
+export * from "./getLastIdentity"
+export * from "./getOrd"
+export * from "./getShow"
+export * from "./partition"
+export * from "./partitionMap"
+export * from "./separate"
+export * from "./separateF"
+export * from "./sequence"
+// codegen:end
