+++ packages/core/src/Option/operations/sequence.ts
@@ -1,0 +1,6 @@
+// ets_tracing: off
+
+import { sequenceF } from "../../Prelude/ForEach"
+import { ForEach } from "../instances/ForEach"
+
+export const sequence = sequenceF(ForEach)
